Move IL db code into instantlaunches/db.go
