+++ instantlaunches/metadata.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/cyverse-de/app-exposer/common"
 	"github.com/labstack/echo/v4"
-	"github.com/lib/pq"
 	"github.com/valyala/fastjson"
 )
 
@@ -81,38 +80,6 @@ func (a *App) listAVUs(c echo.Context) ([]byte, *http.Response, error) {
 	return body, resp, nil
 }
 
-const fullListingQuery = `
-SELECT
-	il.id,
-	ilu.username AS added_by,
-	il.added_on,
-	il.quick_launch_id,
-	ql.name AS ql_name,
-	ql.description AS ql_description,
-	qlu.username AS ql_creator,
-	sub.submission AS submission,
-	ql.app_id,
-	ql.is_public,
-	a.name AS app_name,
-	a.description AS app_description,
-	a.deleted AS app_deleted,
-	a.disabled AS app_disabled,
-	iu.username as integrator
-
-
-FROM instant_launches il
-	JOIN quick_launches ql ON il.quick_launch_id = ql.id
-	JOIN submissions sub ON ql.submission_id = sub.id
-	JOIN apps a ON ql.app_id = a.id
-	JOIN integration_data integ ON a.integration_data_id = integ.id
-	JOIN users iu ON integ.user_id = iu.id
-	JOIN users qlu ON ql.creator = qlu.id
-	JOIN users ilu ON il.added_by = ilu.id
-
-
-WHERE il.id = any($1);
-`
-
 // FullListMetadataHandler returns a list of instant launches with the quick launches
 // embedded, meaning that the submission field is included.
 func (a *App) FullListMetadataHandler(c echo.Context) error {
@@ -140,8 +107,8 @@ func (a *App) FullListMetadataHandler(c echo.Context) error {
 		targetIDs = append(targetIDs, string(avu.GetStringBytes("target_id")))
 	}
 
-	fullListing := []FullInstantLaunch{}
-	if err = a.DB.Select(&fullListing, fullListingQuery, pq.Array(targetIDs)); err != nil {
+	fullListing, err := a.ListFullInstantLaunchesByIDs(targetIDs)
+	if err != nil {
 		if err == sql.ErrNoRows {
 			return echo.NewHTTPError(http.StatusNotFound, "no instant launches found")
 		}
